refactor example app
