+++ example/pubspec.yaml
@@ -7,6 +7,7 @@ dependencies:
   crypto: ^2.1.4
   flutter:
     sdk: flutter
+  file_picker: ^1.13.0
   path_provider: ^1.6.11
   image_picker: ^0.6.7+4
   shared_preferences: ^0.5.8
