+++ ios/Classes/SimpleStreamHandler.swift
@@ -10,6 +10,12 @@ import Foundation
 class SimpleStreamHandler : NSObject, FlutterStreamHandler {
     var cache: [Any?] = []
     
+    let onListen: ((SimpleStreamHandler) -> Void)?
+    
+    init(onListen: ((SimpleStreamHandler) -> Void)? = nil) {
+        self.onListen = onListen
+    }
+    
     var eventSink: FlutterEventSink? = nil
     
     func add(_ value: Any?) {
@@ -22,6 +28,8 @@ class SimpleStreamHandler : NSObject, FlutterStreamHandler {
     }
     
     func onListen(withArguments arguments: Any?, eventSink events: @escaping FlutterEventSink) -> FlutterError? {
+        onListen?(self)
+        
         if !cache.isEmpty {
             for value in cache {
                 events(value)
